Updated code under curation_views from views auth
