+++ src/curation_views.py
@@ -16,8 +16,9 @@ import traceback
 import transaction
 import json
 import re
+from bs4 import BeautifulSoup
 
-from .helpers import allowed_file, extract_id_request, secure_save_file, curator_or_none, extract_references, extract_keywords, get_or_create_filepath, extract_topic, extract_format, file_already_uploaded, link_references_to_file, link_keywords_to_file, FILE_EXTENSIONS, get_locus_by_id, get_go_by_id
+from .helpers import allowed_file, extract_id_request, secure_save_file, curator_or_none, extract_references, extract_keywords, get_or_create_filepath, extract_topic, extract_format, file_already_uploaded, link_references_to_file, link_keywords_to_file, FILE_EXTENSIONS, get_locus_by_id, get_go_by_id,send_newsletter_email
 from .curation_helpers import ban_from_cache, process_pmid_list, get_curator_session, get_pusher_client, validate_orcid
 from .loading.promote_reference_triage import add_paper
 from .models import DBSession, Dbentity, Dbuser, CuratorActivity, Colleague, Colleaguetriage, LocusnoteReference, Referencedbentity, Reservedname, ReservednameTriage, Straindbentity, Literatureannotation, Referencetriage, Referencedeleted, Locusdbentity, CurationReference, Locussummary, validate_tags, convert_space_separated_pmids_to_list
@@ -951,3 +952,74 @@ def get_username_from_db_uri():
 #     transaction.commit() # this commit must be synchronous because the upload_to_s3 task expects the row in the DB
 #     log.info('File ' + request.POST.get('display_name') + ' was successfully uploaded.')
 #     return Response({'success': True})
+
+
+
+@view_config(route_name='colleague_with_subscription', renderer='json', request_method='GET')
+@authenticate
+def colleague_with_subscription(request):
+    colleagues = models_helper.get_all_colleague_with_subscription()
+    return {'colleagues':[colleague.to_simple_dict() for colleague in colleagues]}
+
+@view_config(route_name='get_newsletter_sourcecode',renderer='json',request_method='POST')
+@authenticate
+def get_newsletter_sourcecode(request):
+    try:
+        from urllib2 import urlopen
+        url = str(request.POST['url'])
+
+        if(url.startswith('https://wiki.yeastgenome.org')):
+            response = urlopen(url)
+            html = response.read()
+            soup = BeautifulSoup(html, 'html.parser')
+            body = soup.find(id='content')
+            body['style']='margin-left:0'
+
+            sitesub = body.find(id='siteSub')
+            if(sitesub):
+                sitesub.decompose()
+            
+            jumptonav = body.find(id='jump-to-nav')
+            if(jumptonav):
+                jumptonav.decompose()
+            
+            printfooter= body.find(class_="printfooter")
+            if(printfooter):
+                printfooter.decompose()
+
+            catlinks = body.find(id='catlinks')
+            if(catlinks):
+                catlinks.decompose()
+            
+            for link in body.find_all(href=re.compile("^#")):
+                link['href'] = url + link['href']
+            for link in body.find_all(href=re.compile("^/")):
+                link['href']="https://wiki.yeastgenome.org" + link['href']
+            for img in body.find_all(src=re.compile("^/")):
+                if(img.has_attr('srcset')):
+                    del img['srcset']
+                    # img['srcset'] = ','.join(map(lambda x: 'https://wiki.yeastgenome.org'+ x,img['srcset'].split(',')))
+                img['src']="https://wiki.yeastgenome.org" + img['src']
+            
+            ##Insert wiki's style sheet
+            #stylesheet = BeautifulSoup("<link rel='stylesheet' href='https://wiki.yeastgenome.org/load.php?debug=false&lang=en&modules=mediawiki.legacy.commonPrint%2Cshared%7Cmediawiki.sectionAnchor%7Cmediawiki.skinning.interface%7Cskins.vector.styles&only=styles&skin=vector'/>","lxml")
+            #body.insert(1,stylesheet.link)
+
+            unsubscribe = BeautifulSoup("<p>Note: If you no longer wish to receive this newsletter, please contact the SGD Help Desk at sgd-helpdesk@lists.stanford.edu .<p>").p
+            body.append(unsubscribe)
+            return {"code":body.prettify()}
+        else:
+            return HTTPBadRequest(body=json.dumps({'error': "URL must be from wiki.yeastgenome.org"}))
+    except:
+        return HTTPBadRequest(body=json.dumps({'error': "Unexpected error"}))
+
+@view_config(route_name='send_newsletter',renderer='json',request_method='POST')
+@authenticate
+def send_newsletter(request):
+    try:    
+        html = request.POST['html']
+        subject = request.POST['subject']
+        recipients = request.POST['recipients'].split(';')
+        return send_newsletter_email(subject,recipients,html)
+    except:
+        return {"error":"Error occured during sending newsletter"}
