+++ src/views.py
@@ -8,7 +8,6 @@ from sqlalchemy.exc import IntegrityError
 from datetime import timedelta
 from primer3 import bindings, designPrimers
 from collections import defaultdict
-from bs4 import BeautifulSoup 
 
 import os
 import re
@@ -19,7 +18,7 @@ import logging
 import json
 
 from .models import DBSession, ESearch, Colleague, Dbentity, Edam, Referencedbentity, ReferenceFile, Referenceauthor, FileKeyword, Keyword, Referencedocument, Chebi, ChebiUrl, PhenotypeannotationCond, Phenotypeannotation, Reservedname, Straindbentity, Literatureannotation, Phenotype, Apo, Go, Referencetriage, Referencedeleted, Locusdbentity, Dataset, DatasetKeyword, Contig, Proteindomain, Ec, Dnasequenceannotation, Straindbentity, Disease, Complexdbentity, Filedbentity, Goslim, So, ApoRelation, GoRelation
-from .helpers import extract_id_request, link_references_to_file, link_keywords_to_file, FILE_EXTENSIONS, get_locus_by_id, get_go_by_id, get_disease_by_id, primer3_parser,send_newsletter_email
+from .helpers import extract_id_request, link_references_to_file, link_keywords_to_file, FILE_EXTENSIONS, get_locus_by_id, get_go_by_id, get_disease_by_id, primer3_parser
 from .search_helpers import build_autocomplete_search_body_request, format_autocomplete_results, build_search_query, build_es_search_body_request, build_es_aggregation_body_request, format_search_results, format_aggregation_results, build_sequence_objects_search_query
 from .models_helpers import ModelsHelper
 from .models import SGD_SOURCE_ID, TAXON_ID
@@ -1311,74 +1310,6 @@ def alignment(request):
         return data
     else:
         return {}
-
-@view_config(route_name='colleague_with_subscription', renderer='json', request_method='GET')
-# @authenticate
-def colleague_with_subscription(request):
-    colleagues = models_helper.get_all_colleague_with_subscription()
-    return {'colleagues':[colleague.to_simple_dict() for colleague in colleagues]}
-
-@view_config(route_name='get_newsletter_sourcecode',renderer='json',request_method='POST')
-def get_newsletter_sourcecode(request):
-    try:
-        from urllib2 import urlopen
-        url = str(request.POST['url'])
-
-        if(url.startswith('https://wiki.yeastgenome.org')):
-            response = urlopen(url)
-            html = response.read()
-            soup = BeautifulSoup(html, 'html.parser')
-            body = soup.find(id='content')
-            body['style']='margin-left:0'
-
-            sitesub = body.find(id='siteSub')
-            if(sitesub):
-                sitesub.decompose()
-            
-            jumptonav = body.find(id='jump-to-nav')
-            if(jumptonav):
-                jumptonav.decompose()
-            
-            printfooter= body.find(class_="printfooter")
-            if(printfooter):
-                printfooter.decompose()
-
-            catlinks = body.find(id='catlinks')
-            if(catlinks):
-                catlinks.decompose()
-            
-            for link in body.find_all(href=re.compile("^#")):
-                link['href'] = url + link['href']
-            for link in body.find_all(href=re.compile("^/")):
-                link['href']="https://wiki.yeastgenome.org" + link['href']
-            for img in body.find_all(src=re.compile("^/")):
-                if(img.has_attr('srcset')):
-                    del img['srcset']
-                    # img['srcset'] = ','.join(map(lambda x: 'https://wiki.yeastgenome.org'+ x,img['srcset'].split(',')))
-                img['src']="https://wiki.yeastgenome.org" + img['src']
-            
-            ##Insert wiki's style sheet
-            #stylesheet = BeautifulSoup("<link rel='stylesheet' href='https://wiki.yeastgenome.org/load.php?debug=false&lang=en&modules=mediawiki.legacy.commonPrint%2Cshared%7Cmediawiki.sectionAnchor%7Cmediawiki.skinning.interface%7Cskins.vector.styles&only=styles&skin=vector'/>","lxml")
-            #body.insert(1,stylesheet.link)
-
-            unsubscribe = BeautifulSoup("<p>Note: If you no longer wish to receive this newsletter, please contact the SGD Help Desk at sgd-helpdesk@lists.stanford.edu .<p>").p
-            body.append(unsubscribe)
-            return {"code":body.prettify()}
-        else:
-            return HTTPBadRequest(body=json.dumps({'error': "URL must be from wiki.yeastgenome.org"}))
-    except:
-        return HTTPBadRequest(body=json.dumps({'error': "Unexpected error"}))
-
-
-@view_config(route_name='send_newsletter',renderer='json',request_method='POST')
-def send_newsletter(request):
-    try:    
-        html = request.POST['html']
-        subject = request.POST['subject']
-        recipients = request.POST['recipients'].split(';')
-        return send_newsletter_email(subject,recipients,html)
-    except:
-        return {"error":"Error occured during sending newsletter"}
 
 # check for basic rad54 response
 @view_config(route_name='healthcheck', renderer='json', request_method='GET')
